Move previewer to changeset_spec.go
